[feat #175] fix Image component warning for use of layout prop
