+++ app/client/src/components/Sidebar.tsx
@@ -51,7 +51,7 @@ export const Sidebar = () => {
     <>
       {/* DESKTOP SIDEBAR  */}
       <div className="bg-white min-w-[290px] w-1/5 h-full hidden lg:block px-5 py-7" style={{ boxShadow:"1px 0px 2px rgba(0, 0, 0, 0.1)", zIndex:1000 }}>
-        <Image onClick={()=>router.push("/")} className="cursor-pointer" alt="Landver" src={"/logo-and-name.svg"} height={38} width={157} /> 
+        <Image priority onClick={()=>router.push("/")} className="cursor-pointer" alt="Landver" src={"/logo-and-name.svg"} height={38} width={157} /> 
         <div className="flex flex-col pt-10 gap-6">
           {
             !!loginStore.userType && OPTIONS[loginStore.userType].map((item, index)=>{
@@ -74,7 +74,7 @@ export const Sidebar = () => {
       {/* MOBILE SIDEBAR  */}
       <div style={{ zIndex:100000, transform:showSidebar?"scalex(1)":"scalex(0)", transformOrigin:"left" }} className=" overflow-hidden lg:hidden fixed bottom-4 left-0 top-4 rounded-xl shadow-md shadow-gray-400 w-[90%] sm:w-[400px] bg-white flex flex-col transition-all">
         <div className="bg-[#F3F4F6] py-4 flex justify-center">
-          <Image onClick={()=>{
+          <Image priority onClick={()=>{
             setShowSidebar(false)
             router.push("/")
           }} className="cursor-pointer" alt="Landver" src={"/logo-and-name.svg"} height={38} width={157} />
